Use a double to store state of LabelledSlider (#288)
* Use a double to store state of LabelledSlider

The LabelledSlider is used to change the depth setting. Some
touchpads (e.g. on macOS) can send dozens or hundreds of scroll
events per second, each with a very small delta. Unless we keep
track of the "true" state of the slider, these deltas will be lost
when they are rounded into the integer depth setting.

Fixes #287

* Move preciseValue range check to setValue

Ensures that preciseValue can't be set to invalid values by
setValue callers.

* fix rounding of preciseValue of LabelledSlider below zero

+++ minutor.cpp
@@ -101,10 +101,10 @@ Minutor::Minutor()
 
   connect(depth, SIGNAL(valueChanged(int)),
           mapview, SLOT(setDepth(int)));
-  connect(mapview, SIGNAL(demandDepthChange(int)),
-          depth, SLOT(changeValue(int)));
-  connect(mapview, SIGNAL(demandDepthValue(int)),
-          depth, SLOT(setValue(int)));
+  connect(mapview, SIGNAL(demandDepthChange(double)),
+          depth, SLOT(changeValue(double)));
+  connect(mapview, SIGNAL(demandDepthValue(double)),
+          depth, SLOT(setValue(double)));
   connect(this, SIGNAL(worldLoaded(bool)),
           mapview, SLOT(setEnabled(bool)));
   connect(this, SIGNAL(worldLoaded(bool)),
